Query sf::Keyboard for modifier key state.
The control, alt, etc. fields on KeyPressed / KeyReleased events tell us
which key was involved in the event.  So, if we get a KeyPressed event
on (say) LControl, event.key.control will be true, and the key code will
be the code for LControl.  If we then get a KeyReleased event on
LControl, event.key.control will *still be true* because LControl was
the released key.

Previously, we interpreted event.key.control, alt, etc to mean "this
modifier key is active", but the above makes it clear that our previous
interpretation was wrong.